+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - More CRD documentation ([#433]).
+- Support for exposing HDFS clusters to clients outside of Kubernetes ([#450]).
+- Helm: support labels in values.yaml ([#460]).
 
 ### Changed
 
@@ -14,6 +16,7 @@ All notable changes to this project will be documented in this file.
 
 ### Removed
 
+- [BREAKING] `.spec.clusterConfig.listenerClass` has been split to `.spec.nameNodes.config.listenerClass` and `.spec.dataNodes.config.listenerClass`, migration will be required when using `external-unstable` ([#450], [#462]).
 - [BREAKING] Removed legacy node selector on roleGroups ([#454]).
 - Change default value of `dfs.ha.nn.not-become-active-in-safemode` from `true` to `false` ([#458]).
 
@@ -22,9 +25,12 @@ All notable changes to this project will be documented in this file.
 - Include hdfs principals `dfs.journalnode.kerberos.principal`, `dfs.namenode.kerberos.principal`
   and `dfs.datanode.kerberos.principal` in the discovery ConfigMap in case Kerberos is enabled ([#451]).
 
-[#454]: https://github.com/stackabletech/hdfs-operator/pull/454
+[#450]: https://github.com/stackabletech/hdfs-operator/pull/450
 [#451]: https://github.com/stackabletech/hdfs-operator/pull/451
+[#454]: https://github.com/stackabletech/hdfs-operator/pull/454
 [#458]: https://github.com/stackabletech/hdfs-operator/pull/458
+[#460]: https://github.com/stackabletech/hdfs-operator/pull/460
+[#462]: https://github.com/stackabletech/hdfs-operator/pull/462
 
 ## [23.11.0] - 2023-11-24